fix(firebase): made some corrections
