+++ projects/core/src/services/lessons.data.service.ts
@@ -1,5 +1,10 @@
 import { Injectable } from "@angular/core";
-import { DataBaseService, Lesson } from "./database.service";
+import { DataBaseService } from "./database.service";
+
+export interface Lesson {
+	name: string;
+	description: string;
+}
 
 @Injectable()
 export class LessonsDataService extends DataBaseService<Lesson> {}
