--- src/app/app.component.scss
@@ -1,17 +1,0 @@
-.custom {
-  width: 50%;
-  text-align: center;
-}
-.formats {
-  width: 100%;
-  padding: 10px;
-}
-.btn {
-  padding: 10px;
-  margin-top: 5px;
-  background-color: #eee;
-  width: 60%;
-}
-.btn:hover {
-  cursor: pointer;
-}
